pyproject.toml
[project]
name = "openclaw-whisper"
version = "0.1.0"
description = "Local Whisper STT service for OpenClaw Feishu bot"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
include = ["openclaw_whisper*"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]
